requirements.txt
pyaudio==0.2.13
numpy==1.24.3
pydub==0.25.1
gtts==2.3.2
torch==2.0.0
openai-whisper==20230314

